// src/services/models/mintOnBridgeResponseDto.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * Sonic SVM ZkBridge API
 * OpenAPI spec version: 0.0.1
 */

export interface MintOnBridgeResponseDto {
  readonly amount: string;
  readonly dst: string;
  readonly dstTxHash: string;
  readonly src: string;
  readonly srcTxHash: string;
}